rewrite "VER_1()" to integers and fix bug
VER_1() didn't work for some reason (this is why I started this branch in the first place).
Completely rewriting the whole encryption-logic from scratch solved the issue.
And now this part of QED only uses integers and booleans for calculations too. No messing with strings. (That hopefully increased the speed of QED as well)
This version didn't get mutch testing (not as the finished product), but I tested every function before putting the script together.
Here goes hoping that this thing works.