fix bug in "QED_hilfsfunktionen.IntToBit()"
The bug was making opening files from the GUI break for some unknown reason. The fix was to redo the function using my current knowledge of python.
The reason for why I do not use bg95ght87's function? I do not want to import something inside of a file that gets imported. I want the file to contain everything that needs to be imported. Importing math multiple times feels wasteful to me.