merge with "main"s "QED_system_1_1_2.py"
In this commit, the latest changes from the "main" branch are adopted into the branch. The changes from "main" include optimizations in "VER_1" and "mix_letter".
In addition I added some typehinting and for example made the "print" statement for testing in `if __name__ == "__main__"` look better.

I also added a note to README, but the README file is not important to the actual code (at the moment), so the changes will remain mostly undocumented.