stop gui from en- & decrypting at the same time
Due to performance issues with decrypting large files (tested on a 84 KB file), the decision was made to not decrypt the data at the same time as encrypting it. The user may now choose which action he would like to perform.
This "performance issue" mostly lies in executing "VER_1()" as this functions runtime increases quadratically with growing input lengths. At least for decrypting.