stop using strings of 1s and 0s for text
Back in the days where this was a prototype for a prototype for a prototype, all binary operations where implemented by comparing strings that only consist of 1s and 0s.
This is very inefficient, so we started using integers for many calculations. This makes many things faster, but integrating the faster varialbe types was never fully completed.
Even with this commit we aren't done integrating, but this is a step closer to the finish.
The biggest change of this commit lies within the GUI (finally emotes and other special characters can be printed 😄), but the way the data gets passed to the system has changed a little as well.
The "text" variables are now either integers or bytearrays in (most?) of the code. "KEY" is the only variable that has not yet been converted to some faster type.